Extract AsyncSFTPManager into file_manager.sftp_manager
Phase 4c-i of the file_manager_window decomposition.

Move the async SFTP backend out of the monolith:

- file_manager/sftp_manager.py — FileEntry (dataclass),
  _MainThreadDispatcher, and AsyncSFTPManager. ~1900 lines.

The class uses three lazy imports for runtime helpers; these now
go through the parent package: ssh_config_utils, config, and
scp_utils. Added top-level imports for `re` and `pathlib` since
those were previously satisfied by the monolith's top-of-file
imports and not picked up by the function-local grep used to
pick the verbatim block.

file_manager_window.py re-imports the moved names from
sshpilot.file_manager so window.py / text_editor.py /
sftp_utils.py / file_manager_integration.py / the manual SFTP
scripts keep working. Net: file_manager_window.py
7360 -> 5471 lines.

FilePane and FileManagerWindow still live in
file_manager_window.py and move in 4c-ii / 4c-iii.

Refs #11
